mac_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
bzip2=1.0.8=h99b78c6_7
ca-certificates=2025.1.31=hf0a4a13_0
certifi=2025.1.31=pypi_0
cloudpickle=3.1.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.13=py310h853098b_0
decorator=5.2.1=pyhd8ed1ab_0
distro=1.9.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
farama-notifications=0.0.4=pypi_0
fonttools=4.56.0=pypi_0
gymnasium=1.1.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.34.0=pyh907856f_0
jedi=0.19.2=pyhd8ed1ab_1
jiter=0.9.0=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=h237132a_0
libcxx=19.1.7=ha82da77_0
libedit=3.1.20250104=pl5321hafb1f1b_0
libffi=3.4.2=h3422bc3_5
liblzma=5.6.4=h39f12f2_0
libsodium=1.0.20=h99b78c6_0
libsqlite=3.49.1=h3f77e49_1
libzlib=1.3.1=h8359307_2
matplotlib=3.10.1=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
ncurses=6.5=h5e97a16_3
nest-asyncio=1.6.0=pyhd8ed1ab_1
numpy=2.2.3=pypi_0
openai=1.66.3=pypi_0
opensimplex=0.4.5.1=pypi_0
openssl=3.4.1=h81ee809_0
packaging=24.2=pyhd8ed1ab_2
parso=0.8.4=pyhd8ed1ab_1
pettingzoo=1.24.3=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=pypi_0
pip=25.0.1=pyh8b19718_0
platformdirs=4.3.6=pyhd8ed1ab_1
prompt-toolkit=3.0.50=pyha770c72_0
psutil=7.0.0=py310h078409c_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.1=pypi_0
python=3.10.16=h870587a_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python_abi=3.10=5_cp310
pyzmq=26.3.0=py310h6d25ecc_0
readline=8.2=h1d1bf99_2
ruamel-yaml=0.18.10=pypi_0
ruamel-yaml-clib=0.2.12=pypi_0
setuptools=75.8.2=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
sniffio=1.3.1=pypi_0
stack_data=0.6.3=pyhd8ed1ab_1
tk=8.6.13=h5083fa2_1
tornado=6.4.2=py310h078409c_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025a=h78e105d_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=pyhd8ed1ab_1
